use suspense for dynamic csrfToken input field
